Convert Kind to a uint8. This should conserve memory and be faster.
Many implementations of std::vector will have optimizations for very
small arrays (e.g. < 32 bytes or so). Since we're usually going to have
only a handful of values (probably < 5), making them 1 byte each should
allow our std::vector to never allocate heap memory.

Overall this is a small optimization, but it makes sense since we're not
likely to ever have more than 256 Kinds.

+++ v8_c_bridge.h
@@ -1,4 +1,5 @@
 #include <stddef.h>
+#include <stdint.h>
 
 #ifndef V8_C_BRIDGE_H
 #define V8_C_BRIDGE_H
@@ -36,6 +37,8 @@ typedef struct {
     Error error_msg;
 } ValueErrorPair;
 
+// NOTE! These values must exactly match the values in kinds.go. Any mismatch
+// will cause kinds to be misreported.
 typedef enum {
     kUndefined,
     kNull,
@@ -89,7 +92,7 @@ typedef enum {
 } ValueKind;
 
 typedef struct {
-    const int* ptr;
+    const uint8_t* ptr;
     size_t len;
 } ValueKinds;
